+++ CHANGELOG.md
@@ -3,7 +3,8 @@ All notable changes to this project will be documented in this file.
 
 ## 0.26.1 – May 4, 2020
 ### Fixed
-- Aggregates on nested array fields failed with `value.substr is not function`. 
+- Aggregates on nested array fields failed with `value.substr is not function`.
+- Slow detector for `MIN` `MAX` aggregates must use a specified field as `order by` to detect fast query. 
 
 ## 0.26.0 – May 2, 2020
 ### New